💄 Design: 화면전환 시 slide 제거, navbar 여백 제거
화면전환 시 slide 제거, 알림창으로 가는 것은 슬라이드 있습니다.
