+++ app/(tabs)/_layout.tsx
@@ -1,32 +1,34 @@
-import { Tabs, useRouter } from "expo-router";
+import { Tabs } from "expo-router";
 import React from "react";
+import { StyleSheet, View } from "react-native";
 
 import { AppNavbar } from "@/components/AppNavbar";
+import { TAB_BAR_HEIGHT } from "@/constants/layout";
 
 export default function TabLayout() {
-  const router = useRouter();
-
   return (
     <Tabs
+      detachInactiveScreens={false}
       tabBar={(props) => {
         const currentRouteName = props.state.routes[props.state.index].name;
 
         return (
-          <AppNavbar
-            activeTabId={currentRouteName}
-            onTabPress={(id) => {
-              if (id === "index") {
-                router.push("/home" as never);
-                return;
-              }
-
-              props.navigation.navigate(id);
-            }}
-          />
+          <View style={styles.tabBar}>
+            <AppNavbar
+              activeTabId={currentRouteName}
+              onTabPress={(id) => {
+                props.navigation.navigate(id);
+              }}
+            />
+          </View>
         );
       }}
       screenOptions={{
         headerShown: false,
+        animation: "none",
+        lazy: false,
+        freezeOnBlur: false,
+        sceneStyle: styles.scene,
       }}
     >
       <Tabs.Screen name="index" />
@@ -36,3 +38,13 @@ export default function TabLayout() {
     </Tabs>
   );
 }
+
+const styles = StyleSheet.create({
+  scene: {
+    backgroundColor: "#FFFFFF",
+  },
+  tabBar: {
+    height: TAB_BAR_HEIGHT,
+    backgroundColor: "#FFFFFF",
+  },
+});
