+++ app/club/_layout.tsx
@@ -2,7 +2,7 @@ import { Stack } from "expo-router";
 
 export default function ClubLayout() {
   return (
-    <Stack screenOptions={{ headerShown: false }}>
+    <Stack screenOptions={{ headerShown: false, animation: "none" }}>
       <Stack.Screen name="home" />
       <Stack.Screen name="create" />
       <Stack.Screen name="detail" />
