+++ constants/layout.ts
@@ -1,0 +1,2 @@
+export const TAB_BAR_HEIGHT = 76;
+export const TAB_SCREEN_BOTTOM_PADDING = 24;
